Guardrail-Service/requirements.txt
# Core Framework
Flask==3.0.3
python-dotenv==1.0.1

# Kafka Integration
kafka-python==2.2.15

# Guardrails & LLM
guardrails-ai==0.5.0
openai==1.12.0

# Utilities
requests==2.31.0
jsonschema==4.25.1

# Logging & Monitoring
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.4
pytest-cov==4.1.0

